Updated "Avatars", paper 15 under 2023 tab.
The current commit change corresponds to the latest version (v2) of the paper "FlashAvatar: High-Fidelity Digital Avatar Rendering at 300FPS” with an updated title "FlashAvatar: High-fidelity Head Avatar with Efficient Gaussian Embedding".

Fixed the link of v1 paper with the correct one. It was mistakenly mixed with an incorrect one, with the paper 14 of GaussianAvatar link.
Code had been provided for paper v2.
Authors remained the same.
Abstract remains as it is.